An Android application for 4wd users to select a desired tyre
An Android application that the user can select tyre type as well as
tyre hieght, which in turn displays a specific tyre with picture and
tyre details for the user. The user is also asked to enter his/her
details for notifications, they are able to skip this process is it has
already been done or the user just doesn't want notifications. The main
screens are also able to rotate 90 degrees if user tilts his/her mobile
device.